Split sign in and sign up flows, as they call different methods and have different flows from the Firebase Auth perspective
