+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/java/com/notes/app/screens/sign_up/SignUpViewModel.kt
@@ -1,7 +1,16 @@
 package com.notes.app.screens.sign_up
 
+import android.util.Log
+import androidx.credentials.Credential
+import androidx.credentials.CustomCredential
+import androidx.credentials.GetCredentialRequest
+import com.google.android.libraries.identity.googleid.GetGoogleIdOption
+import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
+import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential.Companion.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL
+import com.notes.app.ERROR_TAG
 import com.notes.app.NOTES_LIST_SCREEN
 import com.notes.app.SIGN_UP_SCREEN
+import com.notes.app.UNEXPECTED_CREDENTIAL
 import com.notes.app.model.service.AccountService
 import com.notes.app.screens.NotesAppViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -36,6 +45,17 @@ class SignUpViewModel @Inject constructor(
         _confirmPassword.value = newConfirmPassword
     }
 
+    fun getCredentialRequest(webClientId: String): GetCredentialRequest {
+        val googleIdOption = GetGoogleIdOption.Builder()
+            .setFilterByAuthorizedAccounts(false)
+            .setServerClientId(webClientId)
+            .build()
+
+        return GetCredentialRequest.Builder()
+            .addCredentialOption(googleIdOption)
+            .build()
+    }
+
     fun onSignUpClick(openAndPopUp: (String, String) -> Unit) {
         launchCatching {
             if (!_email.value.isValidEmail()) {
@@ -52,6 +72,18 @@ class SignUpViewModel @Inject constructor(
 
             accountService.linkAccountWithEmail(_email.value, _password.value)
             openAndPopUp(NOTES_LIST_SCREEN, SIGN_UP_SCREEN)
+        }
+    }
+
+    fun onSignUpWithGoogle(credential: Credential, openAndPopUp: (String, String) -> Unit) {
+        launchCatching {
+            if (credential is CustomCredential && credential.type == TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
+                val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
+                accountService.linkAccountWithGoogle(googleIdTokenCredential.idToken)
+                openAndPopUp(NOTES_LIST_SCREEN, SIGN_UP_SCREEN)
+            } else {
+                Log.e(ERROR_TAG, UNEXPECTED_CREDENTIAL)
+            }
         }
     }
 }
