+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/res/values/strings.xml
@@ -5,9 +5,10 @@
     <string name="password">Password</string>
     <string name="confirm_password">Confirm password</string>
     <string name="sign_in">Sign in</string>
-    <string name="authenticate_with_google">Authenticate with Google</string>
-    <string name="authentication_error">No accounts found on this device</string>
+    <string name="sign_in_with_google">Sign in with Google</string>
+    <string name="authentication_error">No Google accounts found on this device</string>
     <string name="sign_up">Sign up</string>
+    <string name="sign_up_with_google">Sign up with Google</string>
     <string name="sign_up_description">Don\'t have an account? Click here to Sign Up!</string>
     <string name="sign_out">Sign out</string>
     <string name="sign_out_title">Sign out?</string>
